+++ targets/TARGET_RENESAS/TARGET_RZ_A1XX/us_ticker.c
@@ -16,10 +16,10 @@
 #include <stddef.h>
 #include "us_ticker_api.h"
 #include "PeripheralNames.h"
-#include "ostm_iodefine.h"
+#include "iodefine.h"
+#include "cmsis.h"
 
 #include "RZ_A1_Init.h"
-#include "MBRZA1H.h"
 #include "vfp_neon_push_pop.h"
 #include "mbed_critical.h"
